Rework the drawer with android support library. change the status bar to transparent and push the drawer to the very front of the whole UI.
