--- app/src/main/java/com/YC2010/jason/myclass/callbacks/NavigationDrawerCallbacks.java
@@ -1,5 +1,0 @@
-package com.YC2010.jason.myclass.callbacks;
-
-public interface NavigationDrawerCallbacks {
-    void onNavigationDrawerItemSelected(int position);
-}
