+++ app/src/main/java/com/YC2010/jason/myclass/ui/fragments/CalendarFragment.java
@@ -36,7 +36,6 @@ public class CalendarFragment extends Fragment implements WeekView.MonthChangeLi
     private static final int TYPE_FIVE_DAY_VIEW = 2;
     private static final int TYPE_WEEK_VIEW = 3;
     private int mWeekViewType = TYPE_FIVE_DAY_VIEW;
-    private NavigationDrawerFragment mNavigationDrawerFragment;
 
     // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
     private static final String ARG_COURSE = "COURSE";
@@ -45,9 +44,6 @@ public class CalendarFragment extends Fragment implements WeekView.MonthChangeLi
     public void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setHasOptionsMenu(true);
-
-        mNavigationDrawerFragment = (NavigationDrawerFragment)
-                getFragmentManager().findFragmentById(R.id.fragment_drawer);
     }
 
     @Override
@@ -76,10 +72,8 @@ public class CalendarFragment extends Fragment implements WeekView.MonthChangeLi
 
     @Override
     public void onCreateOptionsMenu(Menu menu, MenuInflater inflater) {
-        if (!mNavigationDrawerFragment.isDrawerOpen()) {
             inflater.inflate(R.menu.calendar_menu, menu);
             super.onCreateOptionsMenu(menu, inflater);
-        }
     }
 
     @Override
