+++ app/src/main/res/layout/drawer_header.xml
@@ -1,0 +1,46 @@
+<?xml version="1.0" encoding="utf-8"?>
+<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+android:layout_width="match_parent"
+android:layout_height="168dp"
+android:background="@drawable/wallpaper"
+android:orientation="vertical"
+    >
+
+<ImageView
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    android:id="@+id/profile_image"
+    android:layout_width="64dp"
+    android:layout_height="64dp"
+    android:src="@drawable/avatar"
+    android:layout_marginTop="40dp"
+    android:layout_alignParentLeft="true"
+    android:layout_alignParentStart="true"
+    android:layout_marginStart="16dp" />
+
+<TextView
+    android:layout_width="wrap_content"
+    android:layout_height="wrap_content"
+    android:text=""
+    android:textSize="14sp"
+    android:textColor="#FFF"
+    android:textStyle="bold"
+    android:gravity="left"
+    android:id="@+id/username"
+    android:layout_above="@+id/email"
+    android:layout_alignLeft="@+id/profile_image"
+    android:layout_alignStart="@+id/profile_image" />
+
+<TextView
+    android:layout_width="wrap_content"
+    android:layout_height="wrap_content"
+    android:text=""
+    android:id="@+id/email"
+    android:gravity="left"
+    android:layout_marginBottom="8dp"
+    android:textSize="14sp"
+    android:textColor="#fff"
+    android:layout_alignParentBottom="true"
+    android:layout_alignLeft="@+id/username"
+    android:layout_alignStart="@+id/username" />
+
+</RelativeLayout>
