--- app/src/main/res/layout/drawer_row.xml
@@ -1,12 +1,0 @@
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_height="48dp"
-    android:layout_width="match_parent"
-    >
-
-    <TextView android:id="@+id/item_name" android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:layout_marginLeft="36dp"
-        android:background="?android:attr/selectableItemBackground" android:drawablePadding="16dp"
-        android:textSize="14sp" android:gravity="center_vertical"
-        android:textColor="@color/myTextPrimaryColor" />
-</LinearLayout>
