--- app/src/main/res/layout/fragment_navigation_drawer.xml
@@ -1,71 +1,0 @@
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:id="@+id/googleDrawer"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:background="@android:color/white">
-
-    <RelativeLayout
-        android:id="@+id/navigationHeader"
-        android:layout_width="match_parent"
-        android:layout_height="168dp"
-        android:background="@drawable/wallpaper"
-        android:paddingTop="24dp">
-
-        <ImageView
-            android:id="@+id/imgAvatar"
-            android:layout_width="64dp"
-            android:layout_height="64dp"
-            android:layout_alignParentLeft="true"
-            android:layout_alignParentStart="true"
-            android:layout_alignParentTop="true"
-            android:layout_marginBottom="8dp"
-            android:layout_marginLeft="16dp"
-            android:layout_marginTop="16dp" />
-
-        <LinearLayout
-            android:layout_width="fill_parent"
-            android:layout_height="56dp"
-            android:layout_alignParentLeft="true"
-            android:layout_alignParentStart="true"
-            android:layout_below="@+id/imgAvatar"
-            android:layout_marginBottom="8dp"
-            android:layout_marginLeft="16dp"
-            android:layout_marginRight="16dp"
-            android:orientation="vertical">
-
-            <TextView
-                android:id="@+id/txtUsername"
-                android:layout_width="match_parent"
-                android:layout_height="0dp"
-                android:layout_weight="1"
-                android:textColor="@android:color/white"
-                android:textSize="14sp"
-                android:textStyle="bold" />
-
-            <TextView
-                android:id="@+id/txtUserEmail"
-                android:layout_width="match_parent"
-                android:layout_height="0dp"
-                android:layout_weight="1"
-                android:textColor="@android:color/white" />
-        </LinearLayout>
-    </RelativeLayout>
-
-    <View
-        android:id="@+id/separator"
-        android:layout_width="match_parent"
-        android:layout_height="1dp"
-        android:layout_below="@+id/navigationHeader"
-        android:layout_marginBottom="8dp"
-        android:background="#1f000000" />
-
-    <android.support.v7.widget.RecyclerView
-        android:id="@+id/drawerList"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:layout_below="@+id/separator"
-        android:background="@color/myDrawerBackground"
-        android:clickable="true"
-        android:scrollbars="vertical" />
-
-</RelativeLayout>
