+++ app/src/main/res/layout/toolbar_default.xml
@@ -1,7 +1,7 @@
-<android.support.v7.widget.Toolbar
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    style="@style/DefaultToolBarStyle"
+<?xml version="1.0" encoding="utf-8"?>
+<android.support.v7.widget.Toolbar xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    android:background="?attr/colorPrimary"
-    android:minHeight="@dimen/abc_action_bar_default_height_material" />
+    android:layout_height="@dimen/abc_action_bar_default_height_material"
+    style="@style/DefaultToolBarStyle"
+    android:background="@color/myPrimaryColor"/>
