+++ app/src/main/res/values-v21/styles.xml
@@ -6,6 +6,8 @@
         <item name="colorPrimary">@color/myPrimaryColor</item>
         <item name="colorPrimaryDark">@color/myPrimaryDarkColor</item>
         <item name="colorAccent">@color/myPrimaryColor</item>
+        <item name="android:statusBarColor">@android:color/transparent</item>
+
         <item name="android:textColorPrimary">@color/myTextPrimaryColor</item>
         <item name="android:navigationBarColor">@color/myNavigationColor</item>
         <item name="android:textCursorDrawable">@null</item>
